+++ server/app/src/state/socketMiddleware.ts
@@ -1,11 +1,17 @@
 import {
   createAction,
+  createAsyncThunk,
   Middleware,
   MiddlewareAPI,
   PayloadAction
 } from "@reduxjs/toolkit";
 import { openbooksApi } from "./api";
-import { deleteHistoryItem } from "./historySlice";
+import {
+  addHistoryItem,
+  deleteHistoryItem,
+  HistoryItem,
+  updateHistoryItem
+} from "./historySlice";
 import {
   ConnectionResponse,
   DownloadResponse,
@@ -18,17 +24,18 @@ import {
 } from "./messages";
 import { addNotification } from "./notificationSlice";
 import {
+  addInFlightDownload,
   removeInFlightDownload,
+  setActiveItem,
   setConnectionState,
-  setSearchResults,
   setUsername
 } from "./stateSlice";
 import { AppDispatch, RootState } from "./store";
 import { displayNotification, downloadFile } from "./util";
-import { connectToServer } from "./connectionSlice";
+import { selectActiveIrcServer } from "./connectionSlice";
 
 // Action to send a websocket message to the server
-export const sendMessage = createAction<RequestMessage>("socket/send_message");
+const sendMessage = createAction<RequestMessage>("socket/send_message");
 
 // Web socket redux middleware.
 // Listens to socket and dispatches handlers.
@@ -114,3 +121,80 @@ const route = (dispatch: AppDispatch, msg: MessageEvent<string>): void => {
   dispatch(addNotification(notification));
   displayNotification(notification);
 };
+
+export const sendDownload = createAsyncThunk(
+  "state/send_download",
+  (book: string, { dispatch }) => {
+    dispatch(addInFlightDownload(book));
+    dispatch(
+      sendMessage({
+        type: MessageType.DOWNLOAD,
+        payload: { book }
+      })
+    );
+  }
+);
+
+// Send a search to the server. Add to query history and set loading.
+export const sendSearch = createAsyncThunk(
+  "state/send_sendSearch",
+  (queryString: string, { dispatch, getState }) => {
+    // Send the books search query to the server
+    dispatch(
+      sendMessage({
+        type: MessageType.SEARCH,
+        payload: {
+          query: queryString
+        }
+      })
+    );
+
+    const timestamp = new Date().getTime();
+    const serverName = (getState() as RootState).connection.selectedServer.name;
+
+    // Add query to item history.
+    dispatch(addHistoryItem({ query: queryString, timestamp, serverName }));
+    dispatch(
+      setActiveItem({ query: queryString, timestamp: timestamp, serverName })
+    );
+  }
+);
+
+export const setSearchResults = createAsyncThunk<
+  Promise<void>,
+  SearchResponse,
+  { dispatch: AppDispatch; state: RootState }
+>(
+  "state/set_search_results",
+  async ({ books, errors }: SearchResponse, { dispatch, getState }) => {
+    const activeItem = getState().state.activeItem;
+    if (activeItem === null) {
+      return;
+    }
+    const updatedItem: HistoryItem = {
+      query: activeItem.query,
+      timestamp: activeItem.timestamp,
+      serverName: activeItem.serverName,
+      results: books,
+      errors: errors
+    };
+
+    dispatch(setActiveItem(updatedItem));
+    dispatch(updateHistoryItem(updatedItem));
+  }
+);
+
+export const connectToServer = createAsyncThunk(
+  "connection/connect_to_server",
+  async (_, { dispatch, getState }) => {
+    const { address, port, channel, enableTLS } = selectActiveIrcServer(
+      getState() as RootState
+    );
+    dispatch(
+      sendMessage({
+        type: MessageType.CONNECT,
+        payload: { address: `${address}:${port}`, channel, enableTLS }
+      })
+    );
+  }
+);