+++ server/app/src/state/stateSlice.ts
@@ -1,8 +1,5 @@
-import { createAsyncThunk, createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { addHistoryItem, HistoryItem, updateHistoryItem } from "./historySlice";
-import { MessageType, SearchResponse } from "./messages";
-import { AppDispatch, RootState } from "./store";
-import { sendMessage } from "./socketMiddleware";
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { HistoryItem } from "./historySlice";
 
 interface AppState {
   // TODO: Need to differentiate websocket connection status from IRC connection status
@@ -54,64 +51,6 @@ const stateSlice = createSlice({
   }
 });
 
-const sendDownload = createAsyncThunk(
-  "state/send_download",
-  (book: string, { dispatch }) => {
-    dispatch(addInFlightDownload(book));
-    dispatch(
-      sendMessage({
-        type: MessageType.DOWNLOAD,
-        payload: { book }
-      })
-    );
-  }
-);
-
-// Send a search to the server. Add to query history and set loading.
-const sendSearch = createAsyncThunk(
-  "state/send_sendSearch",
-  (queryString: string, { dispatch }) => {
-    // Send the books search query to the server
-    dispatch(
-      sendMessage({
-        type: MessageType.SEARCH,
-        payload: {
-          query: queryString
-        }
-      })
-    );
-
-    const timestamp = new Date().getTime();
-
-    // Add query to item history.
-    dispatch(addHistoryItem({ query: queryString, timestamp }));
-    dispatch(setActiveItem({ query: queryString, timestamp: timestamp }));
-  }
-);
-
-const setSearchResults = createAsyncThunk<
-  Promise<void>,
-  SearchResponse,
-  { dispatch: AppDispatch; state: RootState }
->(
-  "state/set_search_results",
-  async ({ books, errors }: SearchResponse, { dispatch, getState }) => {
-    const activeItem = getState().state.activeItem;
-    if (activeItem === null) {
-      return;
-    }
-    const updatedItem: HistoryItem = {
-      query: activeItem.query,
-      timestamp: activeItem.timestamp,
-      results: books,
-      errors: errors
-    };
-
-    dispatch(setActiveItem(updatedItem));
-    dispatch(updateHistoryItem(updatedItem));
-  }
-);
-
 export const {
   setActiveItem,
   setConnectionState,
@@ -120,7 +59,5 @@ export const {
   removeInFlightDownload,
   toggleSidebar
 } = stateSlice.actions;
-
-export { stateSlice, sendMessage, sendDownload, sendSearch, setSearchResults };
 
 export default stateSlice.reducer;